.cursor/orchestrator/requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
openai>=1.0.0
pathlib>=1.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
mcp>=0.1.0
python-dotenv>=1.0.0
